Add person and book storage files and modify app file
